Made QueryIndex an interface implemented by a WritableQueryIndex and ReadonlyQueryIndex classes.
