+++ lucene/monitor/src/java/org/apache/lucene/monitor/TermsHashBuilder.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.lucene.monitor;
+
+import java.io.IOException;
+import java.util.Map;
+import org.apache.lucene.index.IndexReader;
+import org.apache.lucene.search.IndexSearcher;
+import org.apache.lucene.search.SearcherFactory;
+
+class TermsHashBuilder extends SearcherFactory {
+  private final Map<IndexReader.CacheKey, WritableQueryIndex.QueryTermFilter> termFilters;
+
+  TermsHashBuilder(Map<IndexReader.CacheKey, WritableQueryIndex.QueryTermFilter> termFilters) {
+    this.termFilters = termFilters;
+  }
+
+  @Override
+  public IndexSearcher newSearcher(IndexReader reader, IndexReader previousReader)
+      throws IOException {
+    IndexSearcher searcher = super.newSearcher(reader, previousReader);
+    searcher.setQueryCache(null);
+    termFilters.put(
+        reader.getReaderCacheHelper().getKey(), new WritableQueryIndex.QueryTermFilter(reader));
+    reader.getReaderCacheHelper().addClosedListener(termFilters::remove);
+    return searcher;
+  }
+}
